add: homepage UI
