+++ src/Components/components.js
@@ -1,2 +1,3 @@
 export { PublicRoute } from "./PublicRoute.jsx";
 export { RequiresAuth } from "./RequiresAuth.jsx";
+export { Header } from "./Header.jsx";
