+++ src/pages/Home.jsx
@@ -1,34 +1,19 @@
-import { NavLink } from "react-router-dom";
+import { Link } from "react-router-dom";
+
 const Home = () => {
-  let activeStyle = {
-    fontWeight: "bold",
-  };
   return (
-    <div className="flex flex-col gap-2 p-3">
-      <NavLink
-        style={({ isActive }) => {
-          return isActive ? activeStyle : undefined;
-        }}
-        to="/"
-      >
-        Home
-      </NavLink>
-      <NavLink
-        style={(isActive) => {
-          isActive ? activeStyle : undefined;
-        }}
-        to="/history"
-      >
-        History
-      </NavLink>
-      <NavLink to="/liked">Liked</NavLink>
-      <NavLink to="/login">Login</NavLink>
-      <NavLink to="/signup">Signup</NavLink>
-      <NavLink to="/playlists">Playlists</NavLink>
-      <NavLink to="/playlists/playlist/abc">Playlist</NavLink>
-      <NavLink to="/videos">Video Listing</NavLink>
-      <NavLink to="/video/1234">Video</NavLink>
-      <NavLink to="/watch-later">WatchLater</NavLink>
+    <div className="flex justify-center items-center h-[calc(100vh-5rem)] bg-[#F0F4F8]">
+      <section className="w-2/4 antialiased flex flex-col gap-5 items-center">
+        <h1 className="font-[roboto] font-black text-[3rem] ">
+          Welcome To Phoenix Video Library
+        </h1>
+        <p className="font-[roboto] font-black text-[1.5rem] ">
+          We provide you all the content to watch and learn Badminton.
+        </p>
+        <Link to="/videos">
+          <button className="btn-md-primary">GET STARTED</button>
+        </Link>
+      </section>
     </div>
   );
 };
